Move --runtimes flag to image subcommand only
Also, update docs to reflect the change.
