+++ docs/docs/references/configuration/cli/trivy_filesystem.md
@@ -59,7 +59,6 @@ trivy filesystem [flags] PATH
       --registry-token string       registry token
       --rekor-url string            [EXPERIMENTAL] address of rekor STL server (default "https://rekor.sigstore.dev")
       --reset                       remove all caches and database
-      --runtime strings             Runtime(s) to use (docker,containerd,podman,remote) (default [docker,containerd,podman,remote])
       --sbom-sources strings        [EXPERIMENTAL] try to retrieve SBOM from the specified sources (oci,rekor)
       --scanners strings            comma-separated list of what security issues to detect (vuln,config,secret,license) (default [vuln,secret])
       --secret-config string        specify a path to config file for secret scanning (default "trivy-secret.yaml")
