+++ docs/docs/references/configuration/cli/trivy_image.md
@@ -81,7 +81,7 @@ trivy image [flags] IMAGE_NAME
       --removed-pkgs                   detect vulnerabilities of removed packages (only for Alpine)
       --report string                  specify a format for the compliance report. (default "summary")
       --reset                          remove all caches and database
-      --runtime strings                Runtime(s) to use (docker,containerd,podman,remote) (default [docker,containerd,podman,remote])
+      --runtimes strings               Runtime(s) to use, in priority order (docker,containerd,podman,remote) (default [docker,containerd,podman,remote])
       --sbom-sources strings           [EXPERIMENTAL] try to retrieve SBOM from the specified sources (oci,rekor)
       --scanners strings               comma-separated list of what security issues to detect (vuln,config,secret,license) (default [vuln,secret])
       --secret-config string           specify a path to config file for secret scanning (default "trivy-secret.yaml")
