+++ docs/docs/references/configuration/cli/trivy_kubernetes.md
@@ -70,7 +70,6 @@ trivy kubernetes [flags] { cluster | all | specific resources like kubectl. eg: 
       --rekor-url string            [EXPERIMENTAL] address of rekor STL server (default "https://rekor.sigstore.dev")
       --report string               specify a report format for the output. (all,summary) (default "all")
       --reset                       remove all caches and database
-      --runtime strings             Runtime(s) to use (docker,containerd,podman,remote) (default [docker,containerd,podman,remote])
       --sbom-sources strings        [EXPERIMENTAL] try to retrieve SBOM from the specified sources (oci,rekor)
       --scanners string             comma-separated list of what security issues to detect (vuln,config,secret,license) (default "vuln,config,secret,rbac")
       --secret-config string        specify a path to config file for secret scanning (default "trivy-secret.yaml")
